requirements.txt
asgiref==3.8.1
##backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.17.1
charset-normalizer==2.1.1
cryptography==44.0.0
dj-rest-auth==7.0.1
Django==4.2.16
django-allauth==65.3.1
django-cors-headers==4.4.0
django-environ==0.9.0
django-filter==24.3
django-quill-editor==0.1.40
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
mssql-django==1.5
pillow==10.4.0
protobuf==4.21.9
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.2.0
pyparsing==3.0.9
pytz==2024.2
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.12